+++ identity-wallet/src/lib.rs
@@ -5,8 +5,6 @@ pub mod state;
 pub mod stronghold;
 pub mod verifiable_credential_record;
 
-use jsonwebtoken::{decode, Algorithm, DecodingKey, Validation};
-
 // Re-exports
 pub use oid4vc::{oid4vc_core, oid4vc_manager, oid4vci, oid4vp, siopv2};
 
@@ -16,13 +14,3 @@ pub use oid4vc::{oid4vc_core, oid4vc_manager, oid4vci, oid4vp, siopv2};
 // The error.rs defines our app_error types, implemented throughout the code using the thiserror crate.
 // The persistence.rs is where we define our app persistence functions.
 // The stronghold.rs is where we implement the stronghold library for our app, which is used to store sensitive data.
-
-/// Get the claims from a JWT without performing validation.
-pub fn get_unverified_jwt_claims(jwt: &serde_json::Value) -> serde_json::Value {
-    let key = DecodingKey::from_secret(&[]);
-    let mut validation = Validation::new(Algorithm::EdDSA);
-    validation.insecure_disable_signature_validation();
-    validation.validate_exp = false;
-    validation.required_spec_claims.clear();
-    decode(jwt.as_str().unwrap(), &key, &validation).unwrap().claims
-}
