+++ identity-wallet/src/state/connections/mod.rs
@@ -3,11 +3,6 @@ pub mod reducers;
 
 use super::FeatTrait;
 
-use oid4vc::{
-    oid4vc_core::authorization_request::{AuthorizationRequest, Object},
-    oid4vp::oid4vp::OID4VP,
-    siopv2::siopv2::SIOPv2,
-};
 use serde::{Deserialize, Serialize};
 use ts_rs::TS;
 
@@ -26,9 +21,3 @@ pub struct Connection {
 
 #[typetag::serde(name = "connection")]
 impl FeatTrait for Connection {}
-
-#[derive(Serialize, Deserialize)]
-pub enum ConnectionRequest {
-    SIOPv2(Box<AuthorizationRequest<Object<SIOPv2>>>),
-    OID4VP(Box<AuthorizationRequest<Object<OID4VP>>>),
-}
