+++ identity-wallet/src/state/connections/reducers/mod.rs
@@ -1,71 +1,1 @@
-pub mod handle_oid4vp_authorization_request;
 pub mod handle_siopv2_authorization_request;
-pub mod read_authorization_request;
-
-use oid4vc::oid4vc_core::authorization_request::{AuthorizationRequest, Object};
-use oid4vc::oid4vp::oid4vp::OID4VP;
-use oid4vc::siopv2::siopv2::SIOPv2;
-
-// Helpers
-
-// TODO: move this functionality to the oid4vc-manager crate.
-pub fn get_siopv2_client_name_and_logo_uri(
-    siopv2_authorization_request: &AuthorizationRequest<Object<SIOPv2>>,
-) -> anyhow::Result<(String, Option<String>, String)> {
-    let connection_url = siopv2_authorization_request
-        .body
-        .redirect_uri
-        .domain()
-        .ok_or(anyhow::anyhow!("unable to get domain from redirect_uri"))?
-        .to_string();
-
-    // Get the client_name and logo_uri from the client_metadata if it exists.
-    Ok(siopv2_authorization_request
-        .body
-        .extension
-        .client_metadata
-        .as_ref()
-        .map(|client_metadata| {
-            let client_name = client_metadata
-                .client_name
-                .as_ref()
-                .cloned()
-                .unwrap_or(connection_url.to_string());
-            let logo_uri = client_metadata.logo_uri.as_ref().map(|logo_uri| logo_uri.to_string());
-
-            (client_name, logo_uri, connection_url.clone())
-        })
-        // Otherwise use the connection_url as the client_name.
-        .unwrap_or((connection_url.to_string(), None, connection_url)))
-}
-
-// TODO: move this functionality to the oid4vc-manager crate.
-pub fn get_oid4vp_client_name_and_logo_uri(
-    oid4vp_authorization_request: &AuthorizationRequest<Object<OID4VP>>,
-) -> anyhow::Result<(String, Option<String>, String)> {
-    let connection_url = oid4vp_authorization_request
-        .body
-        .redirect_uri
-        .domain()
-        .ok_or(anyhow::anyhow!("unable to get domain from redirect_uri"))?
-        .to_string();
-
-    // Get the client_name and logo_uri from the client_metadata if it exists.
-    Ok(oid4vp_authorization_request
-        .body
-        .extension
-        .client_metadata
-        .as_ref()
-        .map(|client_metadata| {
-            let client_name = client_metadata
-                .client_name
-                .as_ref()
-                .cloned()
-                .unwrap_or(connection_url.to_string());
-            let logo_uri = client_metadata.logo_uri.as_ref().map(|logo_uri| logo_uri.to_string());
-
-            (client_name, logo_uri, connection_url.clone())
-        })
-        // Otherwise use the connection_url as the client_name.
-        .unwrap_or((connection_url.to_string(), None, connection_url)))
-}
