+++ identity-wallet/src/state/core_utils/helpers.rs
@@ -1,0 +1,11 @@
+use jsonwebtoken::{decode, Algorithm, DecodingKey, Validation};
+
+/// Get the claims from a JWT without performing validation.
+pub fn get_unverified_jwt_claims(jwt: &serde_json::Value) -> serde_json::Value {
+    let key = DecodingKey::from_secret(&[]);
+    let mut validation = Validation::new(Algorithm::EdDSA);
+    validation.insecure_disable_signature_validation();
+    validation.validate_exp = false;
+    validation.required_spec_claims.clear();
+    decode(jwt.as_str().unwrap(), &key, &validation).unwrap().claims
+}
