+++ identity-wallet/src/state/credentials/actions/credentials_selected.rs
@@ -1,5 +1,5 @@
 use crate::reducer;
-use crate::state::connections::reducers::handle_oid4vp_authorization_request::handle_oid4vp_authorization_request;
+use crate::state::credentials::reducers::handle_oid4vp_authorization_request::handle_oid4vp_authorization_request;
 use crate::state::{actions::ActionTrait, Reducer};
 
 use serde::{Deserialize, Serialize};
