+++ identity-wallet/src/state/credentials/reducers/handle_oid4vp_authorization_request.rs
@@ -1,15 +1,12 @@
 use crate::{
     error::AppError::{self, *},
-    get_unverified_jwt_claims,
     state::{
-        actions::{listen, Action},
-        connections::{reducers::get_oid4vp_client_name_and_logo_uri, Connection, ConnectionRequest},
-        credentials::actions::credentials_selected::CredentialsSelected,
-        user_prompt::CurrentUserPrompt,
-        AppState,
+        actions::{listen, Action}, connections::Connection, core_utils::ConnectionRequest, credentials::actions::credentials_selected::CredentialsSelected, core_utils::helpers::get_unverified_jwt_claims, user_prompt::CurrentUserPrompt, AppState
     },
 };
 
+use oid4vc::oid4vc_core::authorization_request::{AuthorizationRequest, Object};
+use oid4vc::oid4vp::oid4vp::OID4VP;
 use identity_credential::{credential::Jwt, presentation::Presentation};
 use log::info;
 use oid4vc::oid4vc_manager::managers::presentation::create_presentation_submission;
@@ -150,4 +147,37 @@ pub async fn handle_oid4vp_authorization_request(state: AppState, action: Action
     }
 
     Ok(state)
+}
+
+// Helper
+
+// TODO: move this functionality to the oid4vc-manager crate.
+pub fn get_oid4vp_client_name_and_logo_uri(
+    oid4vp_authorization_request: &AuthorizationRequest<Object<OID4VP>>,
+) -> anyhow::Result<(String, Option<String>, String)> {
+    let connection_url = oid4vp_authorization_request
+        .body
+        .redirect_uri
+        .domain()
+        .ok_or(anyhow::anyhow!("unable to get domain from redirect_uri"))?
+        .to_string();
+
+    // Get the client_name and logo_uri from the client_metadata if it exists.
+    Ok(oid4vp_authorization_request
+        .body
+        .extension
+        .client_metadata
+        .as_ref()
+        .map(|client_metadata| {
+            let client_name = client_metadata
+                .client_name
+                .as_ref()
+                .cloned()
+                .unwrap_or(connection_url.to_string());
+            let logo_uri = client_metadata.logo_uri.as_ref().map(|logo_uri| logo_uri.to_string());
+
+            (client_name, logo_uri, connection_url.clone())
+        })
+        // Otherwise use the connection_url as the client_name.
+        .unwrap_or((connection_url.to_string(), None, connection_url)))
 }
