+++ identity-wallet/src/state/credentials/reducers/mod.rs
@@ -1,3 +1,3 @@
-pub mod read_credential_offer;
+pub mod handle_oid4vp_authorization_request;
 pub mod send_credential_request;
 pub mod update_credential_metadata;
