+++ identity-wallet/src/state/qr_code/actions/mod.rs
