+++ identity-wallet/src/state/qr_code/mod.rs
@@ -1,0 +1,2 @@
+pub mod actions;
+pub mod reducers;
