+++ identity-wallet/src/state/qr_code/reducers/mod.rs
@@ -1,0 +1,2 @@
+pub mod read_authorization_request;
+pub mod read_credential_offer;
