--- identity-wallet/src/state/shared/mod.rs
@@ -1,5 +1,0 @@
-pub mod actions;
-pub mod core_utils;
-
-// In this folder the redux pattern will not necessarily be one on one.
-// This folder is meant for shared pieces of code that don't belong to one specific feature.
