+++ identity-wallet/src/state/user_journey/reducers/cancel_user_journey.rs
